# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Bara Sira API'
  description: "Documentation complète de l'API de Bara Sira avec Swagger"
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Serveur principal API'
paths:
  /api/auth/google/redirect:
    get:
      tags:
        - Auth
      summary: 'L5-Swagger endpoint annotations.'
      description: "Controllers stay focused on HTTP behaviour; this class is the single source\nfor operation metadata and references the shared schemas in Schemas.php."
      operationId: google.redirect
      parameters:
        -
          name: intent
          in: query
          description: 'SSO entry point'
          schema:
            type: string
            default: login
            enum:
              - login
              - register
      responses:
        302:
          description: 'Redirect to Google'
      security: []
  /api/auth/google/callback:
    get:
      tags:
        - Auth
      description: 'Authenticates an existing account by Google email. Unknown users are not created and are redirected to registration.'
      operationId: google.callback
      parameters:
        -
          name: code
          in: query
          schema:
            type: string
        -
          name: state
          in: query
          schema:
            type: string
      responses:
        302:
          description: 'Dashboard or registration redirect'
      security: []
  /api/login:
    post:
      tags:
        - Auth
      operationId: auth.login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        422:
          $ref: '#/components/responses/ValidationError'
      security: []
  /api/register:
    post:
      tags:
        - Auth
      operationId: auth.register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        422:
          $ref: '#/components/responses/ValidationError'
      security: []
  /api/logout:
    post:
      tags:
        - Auth
      operationId: auth.logout
      responses:
        200:
          description: 'Logged out'
        401:
          $ref: '#/components/responses/Unauthenticated'
      security:
        -
          sanctum: []
  /api/me:
    get:
      tags:
        - Auth
      operationId: auth.me
      responses:
        200:
          description: 'Current user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthenticated'
      security:
        -
          sanctum: []
  /api/users:
    get:
      tags:
        - Users
      operationId: users.index
      responses:
        200:
          description: Users
        403:
          $ref: '#/components/responses/Forbidden'
      security:
        -
          sanctum: []
  '/api/users/{id}':
    get:
      tags:
        - Users
      operationId: users.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
    put:
      tags:
        - Users
      operationId: users.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    post:
      tags:
        - Users
      operationId: users.updatePost
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
    patch:
      tags:
        - Users
      operationId: users.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
  /api/service-categories:
    get:
      tags:
        - 'Service categories'
      operationId: categories.index
      responses:
        200:
          description: Categories
      security: []
    post:
      tags:
        - 'Service categories'
      operationId: categories.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategoryRequest'
      responses:
        201:
          description: Created
        422:
          $ref: '#/components/responses/ValidationError'
      security: []
  '/api/service-categories/{id}':
    get:
      tags:
        - 'Service categories'
      operationId: categories.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Category
      security: []
    put:
      tags:
        - 'Service categories'
      operationId: categories.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategoryRequest'
      responses:
        200:
          description: Updated
      security: []
    delete:
      tags:
        - 'Service categories'
      operationId: categories.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - 'Service categories'
      operationId: categories.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCategoryRequest'
      responses:
        200:
          description: Updated
      security: []
  /api/services:
    get:
      tags:
        - Services
      operationId: services.index
      responses:
        200:
          description: Services
      security: []
    post:
      tags:
        - Services
      operationId: services.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        201:
          description: Created
      security:
        -
          sanctum: []
  /api/services-search:
    get:
      tags:
        - Services
      operationId: services.search
      parameters:
        -
          name: query
          in: query
          schema:
            type: string
      responses:
        200:
          description: 'Search results'
      security: []
  '/api/services/{id}':
    get:
      tags:
        - Services
      operationId: services.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Service
      security: []
    put:
      tags:
        - Services
      operationId: services.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Services
      operationId: services.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security:
        -
          sanctum: []
    patch:
      tags:
        - Services
      operationId: services.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
  /api/missions:
    get:
      tags:
        - Missions
      operationId: missions.index
      responses:
        200:
          description: Missions
      security:
        -
          sanctum: []
    post:
      tags:
        - Missions
      operationId: missions.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionRequest'
      responses:
        201:
          description: Created
      security:
        -
          sanctum: []
  '/api/missions/{id}':
    get:
      tags:
        - Missions
      operationId: missions.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Mission
      security:
        -
          sanctum: []
    put:
      tags:
        - Missions
      operationId: missions.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
    delete:
      tags:
        - Missions
      operationId: missions.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security:
        -
          sanctum: []
    patch:
      tags:
        - Missions
      operationId: missions.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdateRequest'
      responses:
        200:
          description: Updated
      security:
        -
          sanctum: []
  '/api/missions/{id}/claim':
    post:
      tags:
        - Missions
      operationId: missions.claim
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Claimed
        409:
          description: 'Schedule conflict'
      security:
        -
          sanctum: []
  /api/missions/generate-with-ai:
    post:
      tags:
        - AI
      operationId: missions.generateAi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionAiRequest'
      responses:
        200:
          description: 'Generated mission'
        429:
          description: 'Rate limited'
      security:
        -
          sanctum: []
  /api/messages:
    get:
      tags:
        - Messages
      operationId: messages.index
      responses:
        200:
          description: Conversations
      security:
        -
          sanctum: []
    post:
      tags:
        - Messages
      operationId: messages.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        201:
          description: Sent
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
  '/api/messages/conversation/{user}':
    get:
      tags:
        - Messages
      operationId: messages.conversation
      parameters:
        -
          name: user
          in: path
          required: true
          schema:
            type: integer
        -
          name: mission_id
          in: query
          schema:
            type: integer
            nullable: true
      responses:
        200:
          description: Conversation
      security:
        -
          sanctum: []
  /api/reviews:
    post:
      tags:
        - Reviews
      operationId: reviews.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        201:
          description: Published
      security:
        -
          sanctum: []
  '/api/reviews/{id}':
    patch:
      tags:
        - Reviews
      operationId: reviews.update
      parameters:
        -
          $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateRequest'
      responses:
        200:
          description: 'Final revision'
      security:
        -
          sanctum: []
  /api/ai/translate:
    post:
      tags:
        - AI
      operationId: ai.translate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequest'
      responses:
        200:
          description: Translation
      security:
        -
          sanctum: []
  /api/user-skills:
    get:
      tags:
        - 'User skills'
      operationId: skills.index
      responses:
        200:
          description: Skills
      security: []
    post:
      tags:
        - 'User skills'
      operationId: skills.store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSkillRequest'
      responses:
        201:
          description: Created
      security: []
  '/api/user-skills/{id}':
    get:
      tags:
        - 'User skills'
      operationId: skills.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: Skill
      security: []
    put:
      tags:
        - 'User skills'
      operationId: skills.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - 'User skills'
      operationId: skills.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - 'User skills'
      operationId: skills.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
...
  '/api/resumes/{id}':
    get:
      tags:
        - Resumes
      operationId: resumes.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    put:
      tags:
        - Resumes
      operationId: resumes.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - Resumes
      operationId: resumes.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    patch:
      tags:
        - Resumes
      operationId: resumes.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
...
  '/api/portfolio-items/{id}':
    get:
      tags:
        - Portfolio
      operationId: portfolio.show
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Portfolio item'
      security: []
    put:
      tags:
        - Portfolio
      operationId: portfolio.updatePut
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
    delete:
      tags:
        - Portfolio
      operationId: portfolio.destroy
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        204:
          description: Deleted
      security: []
    patch:
      tags:
        - Portfolio
      operationId: portfolio.updatePatch
      parameters:
        -
          $ref: '#/components/parameters/Id'
      responses:
        200:
          description: 'Not implemented'
      security: []
  /api/payments/webhooks/cinetpay:
    get:
      tags:
        - Payments
      operationId: payments.cinetpayPing
      responses:
        204:
          description: 'Webhook availability'
      security: []
    post:
      tags:
        - Payments
      operationId: payments.cinetpayWebhook
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                cpm_trans_id:
                  type: string
              type: object
      responses:
        204:
          description: 'Notification verified'
      security: []
components:
  schemas:
    User:
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - client
            - prestataire
            - admin
            - superadmin
        avatar_url:
          type: string
          nullable: true
      type: object
    AuthResponse:
      properties:
        success:
          type: boolean
        message:
          type: string
        redirect:
          type: string
        user:
          $ref: '#/components/schemas/User'
      type: object
    LoginRequest:
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        remember:
          type: boolean
      type: object
    RegisterRequest:
      required:
        - first_name
        - last_name
        - phone
        - email
        - password
        - role
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        role:
          type: string
          enum:
            - client
            - prestataire
            - admin
            - superadmin
      type: object
    UserUpdateRequest:
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        bio:
          type: string
          maxLength: 2000
          nullable: true
        avatar:
          type: string
          format: binary
          nullable: true
        remove_avatar:
          type: boolean
      type: object
    ServiceCategoryRequest:
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 150
        description:
          type: string
          nullable: true
        icon:
          type: string
          maxLength: 255
          nullable: true
      type: object
    ServiceRequest:
      required:
        - service_category_id
        - name
        - description
        - price_min
        - price_max
      properties:
        service_category_id:
          type: integer
        name:
          type: string
          maxLength: 150
        description:
          type: string
        icon:
          type: string
          nullable: true
        price_min:
          type: number
          minimum: 0
        price_max:
          type: number
          minimum: 0
        is_active:
          type: boolean
      type: object
    MissionRequest:
      required:
        - service_id
        - title
        - description
        - city
        - address
        - date_start
      properties:
        service_id:
          type: integer
        title:
          type: string
          maxLength: 255
        description:
          type: string
        city:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 255
        skills:
          type: array
          items:
            type: string
          maxItems: 10
        questions:
          type: array
          items:
            type: string
          maxItems: 5
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        price:
          type: number
          minimum: 0
          nullable: true
        date_start:
          type: string
          format: date-time
        date_end:
          type: string
          format: date-time
          nullable: true
      type: object
    MissionUpdateRequest:
      allOf:
        -
          $ref: '#/components/schemas/MissionRequest'
        -
          properties:
            status:
              type: string
              enum:
                - pending
                - in_progress
                - completed
                - cancelled
          type: object
    MissionAiRequest:
      required:
        - keywords
      properties:
        keywords:
          type: string
          maxLength: 1000
          minLength: 3
      type: object
    MessageRequest:
      required:
        - receiver_id
        - message
      properties:
        receiver_id:
          type: integer
        mission_id:
          type: integer
          nullable: true
        message:
          type: string
          maxLength: 5000
      type: object
    ReviewRequest:
      required:
        - mission_id
        - rating
      properties:
        mission_id:
          type: integer
        rating:
          type: integer
          maximum: 5
          minimum: 1
        comment:
          type: string
          maxLength: 2000
          nullable: true
      type: object
    ReviewUpdateRequest:
      required:
        - rating
      properties:
        rating:
          type: integer
          maximum: 5
          minimum: 1
        comment:
          type: string
          maxLength: 2000
          nullable: true
      type: object
    TranslationRequest:
      required:
        - text
        - source_locale
        - target_locale
      properties:
        text:
          type: string
          maxLength: 10000
        source_locale:
          type: string
          maxLength: 10
        target_locale:
          type: string
          maxLength: 10
        context:
          type: object
      type: object
    UserSkillRequest:
      required:
        - user_id
        - service_id
        - level
      properties:
        user_id:
          type: integer
        service_id:
          type: integer
        level:
          type: string
          maxLength: 50
      type: object
  responses:
    Unauthenticated:
      description: 'Authentication required'
    Forbidden:
      description: 'Action forbidden'
    NotFound:
      description: 'Resource not found'
    ValidationError:
      description: 'Validation error'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
              errors:
                type: object
            type: object
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Laravel Sanctum session cookie'
      name: laravel_session
      in: cookie
...